refactor: add is_number_symbol function
Add the is_number_symbol function to the symbol module. This function uses regular expressions to check if a given string represents a number symbol. It improves the functionality of the module by providing a convenient way to determine if a symbol is a number.

Refactor the symbol.h file to include the <stdbool.h> header for the bool type definition.

Refactor the symbol.c file to include the <regex.h> header for regular expression functionality.